pyproject.toml
[tool.poetry]
name = "builder"
version = "0.1.0"
description = ""
authors = ["William <wosteimer1@gmail.com>"]
readme = "README.md"
packages = [{ include = "builder", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
typing-extensions = "^4.11.0"
libsass = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-sugar = "^1.0.0"
isort = "^5.13.2"
black = "^24.4.2"

[tool.pyright]
include = ["src"]
typeCheckingMode = "strict"
enableExperimentalFeatures = true

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
